commandline
command line args processing added
moved launch code into separate class

getting close to prime time
