+++ VS/RBP Launcher/Launcher_Headless/Program.cs
@@ -1,12 +1,15 @@
 ﻿// See https://aka.ms/new-console-template for more information
 using Launcher_Headless.Utilities;
 using Serilog;
+using Microsoft.Extensions.Configuration;
 
 // setup logger
 LoggingToFile.SetupFileLogger();
 
 try
 {
+    // get seetings file path from args past in
+    string filePathFlow = GetFlowSettingsFromArgs(args);
     // read app settings
     var appSettings = RBP_Launcher.Utilities.Configs.AppSettings.GetAppSettings();
     if (appSettings != null)
@@ -15,66 +18,13 @@ try
         // TODO: get settings path from arguments past in
 
         // Replace "path/to/your/file.json" with the actual path to your JSON file
-        string filePathFlow = @"C:\Users\janchristel\Documents\GitHub\SampleCodeRevitBatchProcessor\VS\RBP Launcher\Launcher_Headless\flowSampleSettings.json";
+        //string filePathFlow = @"C:\Users\janchristel\Documents\GitHub\SampleCodeRevitBatchProcessor\VS\RBP Launcher\Launcher_Headless\flowSampleSettings.json";
         Console.WriteLine($"Reading flow settings from {filePathFlow}");
         var scriptConfig = RBP_Launcher.Utilities.Configs.FlowSettings.GetFlowSettings(filePathFlow);
         if (scriptConfig != null)
         {
-            // start excuting flow
-            Console.WriteLine("Starting ...");
-            Log.Information("Starting ...");
-            // get python script runners
-            Dictionary<string, RBP_Launcher.IScriptRunner> pythonScriptRunners = RBP_Launcher.Utilities.PythonScriptRunners.GetAvailablePythonScriptRunners();
-            // run pre flow scripts
-            Console.WriteLine("Starting pre flow groups scripts...");
-            Log.Information("Starting pre flow groups scripts...");
-            bool preScriptsExecutionStatus = Launcher_Headless.Utilities.ExcuteScripts.RunScripts(scriptConfig.PreScript, pythonScriptRunners);
-            // check if all scripts finished successfully
-            if (preScriptsExecutionStatus)
-            {
-                int counter = 0;
-                Log.Debug("All pre flow scripts executed successfully.");
-                Console.WriteLine("Batch processor flow session...");
-                Log.Information("Batch processor flow session...");
-                // start all scrit groups
-                foreach (RBP_Launcher.Utilities.Configs.ScriptConfiguration.Script rbpScriptGroup in scriptConfig.BatchProcessorScripts)
-                {
-                    counter++;
-                    
-                    //log out put
-                    Console.WriteLine($"Starting batch processor flow session {counter} of {scriptConfig.BatchProcessorScripts.Count}");
-                    Log.Information($"Starting batch processor flow session {counter} of {scriptConfig.BatchProcessorScripts.Count}");
-
-                    bool flowGroupExecutionStatus = Launcher_Headless.Utilities.ExcuteScripts.RunBatchProcessorScripts(
-                        rbpScriptGroup,
-                        appSettings,
-                        pythonScriptRunners);
-
-                    if (!flowGroupExecutionStatus)
-                    {
-                        throw new Exception($"Flow group {counter} failed to execute without an exception. Exiting.");
-                    }
-                }
-
-                // run post flow scripts
-                Console.WriteLine("Starting post flow groups scripts...");
-                Log.Information("Starting post flow groups scripts...");
-                bool postScriptsExecutionStatus = Launcher_Headless.Utilities.ExcuteScripts.RunScripts(scriptConfig.PostScript, pythonScriptRunners);
-                // check if all scripts finished successfully
-                if (postScriptsExecutionStatus)
-                {
-                    Log.Debug("All post flow group scripts executed successfully.");
-                    Console.WriteLine("Finished!");
-                }
-                else
-                {
-                    throw new Exception("One or multiple post flow scripts failed to execute. Exiting");
-                }
-            }
-            else
-            {
-                throw new Exception("One or multiple pre flow scripts failed to execute. Exiting");
-            }
+            //start launcher ...
+            RBP_Launcher.Launcher.LaunchIt(appSettings, scriptConfig);
         }
         else
         {
@@ -96,3 +46,35 @@ catch(Exception ex)
 // Close and flush the log when the application exits
 Log.CloseAndFlush();
 
+/// <summary>
+/// Retrieves --settings value from arguments
+/// if not arguments are supplied an exception is thrown
+/// </summary>
+static string GetFlowSettingsFromArgs(string[] args)
+{
+    var configBuilder = new ConfigurationBuilder()
+        .AddCommandLine(args); // This line adds command line arguments to configuration
+
+    var configuration = configBuilder.Build();
+
+    // YAccess configuration value for settings
+    string filePathFlow = configuration["settings"];
+    if (filePathFlow == null)
+    {
+        PrintHelpText();
+        throw new Exception("No --settings argument provided. Exiting");
+    }
+    Log.Information($"Value of --settings: {filePathFlow}");
+    return filePathFlow;
+}
+
+/// <summary>
+/// Prints out a help text on how to supply arguments 
+/// </summary>
+static void PrintHelpText()
+{
+    Console.WriteLine("Usage: Launcher_Headless.exe --settings <file path>");
+    Console.WriteLine("Description: ");
+    Console.WriteLine(" --settings \t\trequires the fully qualified file path to the flow settings .json file");
+    Console.WriteLine("\t\t\tsample: Launcher_Headless.exe --settings \"C/my/file/samle.json\"");
+}
